requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
requests==2.31.0

